// src/lib/pishop/types.ts
export type Maybe<T> = T | null

export type Cart = {
  id: string | undefined
  checkoutUrl: string
  cost: {
    subtotalAmount: Money
    totalAmount: Money
    totalTaxAmount: Money
  }
  totalQuantity: number
  lines: CartItem[]
}

export type CartItem = {
  id: string | undefined
  quantity: number
  cost: {
    totalAmount: Money
  }
  merchandise: {
    id: string
    title: string
    selectedOptions: {
      name: string
      value: string
    }[]
    product: Pick<Product, 'id' | 'handle' | 'title' | 'featuredImage'>
  }
}

export type Collection = {
  handle: string
  title: string
  path: string
  description: string
  seo: SEO
  updatedAt: string
}

export type Image = {
  url: string
  width: number
  height: number
  altText: string
}

export type Menu = {
  title: string
  path: string
}

export type Money = {
  amount: string
  currencyCode: string
}

export type Page = {
  id: string
  title: string
  handle: string
  body: string
  bodySummary: string
  seo: SEO
  createdAt: string
  updatedAt: string
}

export type Product = {
  id: string
  handle: string
  availableForSale: boolean
  title: string
  description: string
  descriptionHtml: string
  // options: ProductOption[];
  options: { [key: string]: string }
  priceRange: {
    maxVariantPrice: Money
    minVariantPrice: Money
  }
  featuredImage: Image
  seo: SEO
  tags: string[]
  updatedAt: string
  variants: ProductVariant[]
  images: Image[]
  rating?: number
  ratingCount?: number
}

export type ProductOption = {
  id: string
  name: string
  values: string[]
}

export type ProductVariant = {
  id: string
  title: string
  availableForSale: boolean
  selectedOptions: {
    name: string
    value: string
  }[]
  price: Money
}

export type SEO = {
  title: string
  description: string
}